extract barcode interface to a single source file
