+++ src/main_debug.cpp
@@ -1,15 +1,14 @@
 //
 // Created by nanos on 2020/10/21.
 //
-#include "detector/detect.hpp"
-#include "decoder/ean_decoder.hpp"
+#include "barcode.hpp"
 
 int main(int argc, char **argv) {
     using namespace cv;
     std::cout << __DATE__ << " " << __TIME__ << std::endl;
     if (argc < 2) {
-        printf("Usage: ./detect_debug --webcam\n"
-               "./detect_debug <input_file>\n");
+        printf("Usage: ./main_debug --webcam\n"
+               "./main_debug <input_file>\n");
         exit(-1);
     }
     BarcodeDetector bardet;
